Add flyover smoke test utility functions and refactor tests to utilize them
